refactor: separate layer state ui from screens
