+++ auth/presentation/src/main/java/com/projectlab/travelin_android/presentation/screens/login/LoginState.kt
@@ -1,0 +1,9 @@
+package com.projectlab.travelin_android.presentation.screens.login
+
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
+
+data class LoginState(
+    val email: MutableState<String> = mutableStateOf(""),
+    val password: MutableState<String> = mutableStateOf("")
+)
