+++ auth/presentation/src/main/java/com/projectlab/travelin_android/presentation/screens/profile/ProfileScreen.kt
@@ -17,7 +17,8 @@ fun ProfileScreen(
             ProfileContent(
                 paddingValues = it,
                 viewModel = viewModel,
-                onLogoutClick = onLogoutClick
+                onLogoutClick = onLogoutClick,
+                state = viewModel.state
             )
         },
         bottomBar = { BottomNavigationBar("profile", onHomeClick) }
