--- auth/presentation/src/main/java/com/projectlab/travelin_android/presentation/screens/profile/components/ProfileUser.kt
@@ -1,79 +1,0 @@
-package com.projectlab.travelin_android.presentation.screens.profile.components
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import com.projectlab.auth.presentation.R
-import com.projectlab.core.presentation.designsystem.theme.spacing
-import com.projectlab.travelin_android.presentation.screens.profile.ProfileViewModel
-
-@Composable
-fun ProfileUser(
-    modifier: Modifier = Modifier,
-    viewModel: ProfileViewModel
-) {
-    val userName = "John Doe"
-    val userBio = "Mars, Solar System"
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MaterialTheme.spacing.ScreenHorizontalPadding),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        Image(
-            painter = painterResource(id = R.drawable.blank_profile_picture),
-            contentDescription = "Profile Photo",
-            modifier = Modifier
-                .size(MaterialTheme.spacing.ImageSize)
-                .clip(CircleShape)
-        )
-        Spacer(modifier = Modifier.width(MaterialTheme.spacing.SectionSpacing))
-
-        Column {
-            Text(
-                text = viewModel.currentUser!!.email!!,
-                style = MaterialTheme.typography.titleMedium.copy(
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-            )
-            Text(
-                text = userBio,
-                style = MaterialTheme.typography.bodyMedium.copy(
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            )
-
-            Text(
-                text = viewModel.userEntityData.email,
-                style = MaterialTheme.typography.titleMedium.copy(
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-            )
-            Text(
-                text = viewModel.userEntityData.age,
-                style = MaterialTheme.typography.titleMedium.copy(
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-            )
-        }
-    }
-}
